Extract fuzzy module to its own crate
We still have a fuzzy module in the zed crate with some app-specific logic.
