+++ fuzzy/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "fuzzy"
+version = "0.1.0"
+edition = "2018"
+
+[dependencies]
