+++ sum_tree/Cargo.toml
@@ -3,8 +3,6 @@ name = "sum_tree"
 version = "0.1.0"
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 arrayvec = "0.7.1"
